[ADD] Paginator limiter [FIX] Service AuthUserSite [FIX] Header localization

+++ servicestation-business/src/main/java/com/expositds/sjc/servicestation/business/repository/dao/OrderDao.java
@@ -15,6 +15,6 @@ public interface OrderDao extends GenericDao<OrderEntity, Long> {
 	
 	public List<OrderEntity> getOrdersStationLimit(Long id, Long first, Long size);
 	
-	public Integer getSiteUserOrdersCount(Long id);
+	public Long getSiteUserOrdersCount(Long id);
 
 }
